The MCTupleWithBackgroundEntailmentModel (#215)
* changed pickle to dill for pickling defaultdicts and added boolean to max lengths lookup

* added start of a lexical overlap analysis script

* added a nounverb filter for question tokens

* added to analysis file

* path change

* path change

* added 2 local todos

* added the Query trait, converted previous string queries to StringQuery and implemented BoostedQuery

* added boost params

* temp added some prints

* fixed typo in param

* returned to original efs paths

* removed debug println

* added some comments to Query

* added .idea/ to gitignore

* added start for a MCTF tuple entailment model

* continued work on tuple match

* framework in place, several todos in tuple_match, and need to write noisy-or/and layers

* added noisy-or start, fixed tupleMatch to add the trainable weights in with build(), added compute_mask() for tupleMatch

* finished noisyOR, pre-test though

* fixed some masking and todos, preventing dividing by 0 in tupleMatch, preventing counting matches with padding

* one more todo removed (prev addressed)

* started to write test for noisy or layer, adding alternate noise parameter initialization

* noisy or tested, compilation test of tuplematch completed

* test written for tuple_match general case

* MultiInputTimeDistributed fixed, tested for general and batch, doesn't currently support masks

* started writing the new MTFTupleWithBackground instance, mostly just with placeholders for now

* oops, added a space to another test, now fixed

* override methods implemented, not yet tested

* tested general case of read_from_line

* tested general case of read_from_line

* MC tuple with background instance is tested

* added script to merge labels into svo tuple data

* swicthed the tuple delimiter to 7215$

* fixed bugs with model

* bugs fixed, now does_not_crash_test passes

* added stemming and filtering punct

* added svo model components

* changes to experiment params

* formatting and in process of moving stemming and filtering to another place

* added an intermediate class (WordProcessor) that adds a layer to the splitting-tokenization pipline for postprocessing

* oops -- forgot these -- removed filtering and stemming from the instance

* fixed mistake in docstring and also added instance to the docs

* added mc_tuples_with_background_instance to doc/data/instances.rst

* updating to use the newer version of timedistributed, also added fix to tuple match output shape and a squeeze layer to wrap K.squeeze

* merging part of completed PRs into svo code, but now not learning/building inner tuple_match layer

* shoot, spacing in word_splitter

* removed comments

* added the TimeDistributed fix that builds the super of Keras' timeDistributed

* added docs and docstrings, cleaned up code

* updated the comments in model to match the new equivalence of num_slots in question tuples and background tuples

* updated tests for the newer delimiters in tuple_inference_instance

* renamed model to better match the instance, etc

* fixing pylint too-long-lines

* removed add_label script

* removed .idea from local gitignore

* renamed model

* renamed the tuple_match params

* note about TupleMatch WordsAndChars tokenizer incompatibility added

* NoisyOr added to custom objects

* calling super in _get_max_lengths, also added optional word_character_length for good measure and potential compatibility

* calling super in _set_max_lengths and also added word_character_length

* removed max from slot_length, also added the _get_sentence_shape for the question and background input

* oops, missed a param rename, fixed

* fixed copy-paste bug

* added params for tiling repetitions, added test for laoding

* added get_configs to squeeze, repeat, and tuple_match, now passing load model test

* final cleanup, removed name param, removed setting params that are set in super class, and consolidated tests

+++ tests/layers/tuple_match_test.py
@@ -15,8 +15,8 @@ class TestTupleMatch(TestCase):
         num_slots = 3
         t1_words_per_slot = 4
         t2_words_per_slot = 3
-        self.tuple1_input = Input(shape=(num_slots, t1_words_per_slot), dtype='float32', name="input_tuple1")
-        self.tuple2_input = Input(shape=(num_slots, t2_words_per_slot), dtype='float32', name="input_tuple2")
+        self.tuple1_input = Input(shape=(num_slots, t1_words_per_slot), dtype='int32', name="input_tuple1")
+        self.tuple2_input = Input(shape=(num_slots, t2_words_per_slot), dtype='int32', name="input_tuple2")
         self.num_hidden_layers = 1
         self.hidden_layer_width = 2
         self.hidden_layer_activation = 'linear'
